requirements.txt
discord
telepot
PyYAML
pyrlottie
Pillow
coloredlogs